Moved main methods for indvidual Runners into common Runner trait
