+++ geomesa-tools/src/main/scala/org/locationtech/geomesa/tools/accumulo/AccumuloRunner.scala
@@ -32,16 +32,4 @@ object AccumuloRunner extends Runner {
     new QueryStatsCommand(jc),
     new GetSftCommand(jc)
   )
-
-  def main(args: Array[String]): Unit = {
-    val command = createCommand(args)
-    try {
-      command.execute()
-    } catch {
-      case e: Exception =>
-        logger.error(e.getMessage, e)
-        sys.exit(-1)
-    }
-    sys.exit(0)
-  }
 }
