+++ geomesa-tools/src/main/scala/org/locationtech/geomesa/tools/common/Runner.scala
@@ -2,9 +2,8 @@ package org.locationtech.geomesa.tools.common
 
 import com.beust.jcommander.{JCommander, ParameterException}
 import com.typesafe.scalalogging.LazyLogging
+import org.locationtech.geomesa.tools.common.commands.Command
 import org.locationtech.geomesa.tools.common.commands.convert.GeoMesaIStringConverterFactory
-import org.locationtech.geomesa.tools.common.commands.{Command, HelpCommand, VersionCommand}
-import org.locationtech.geomesa.tools.kafka.commands.CreateCommand
 
 import scala.collection.JavaConversions._
 
@@ -12,6 +11,18 @@ trait Runner extends LazyLogging {
   val jc = new JCommander()
   val scriptName: String
   val commands: List[Command]
+
+  def main(args: Array[String]): Unit = {
+    val command = createCommand(args)
+    try {
+      command.execute()
+    } catch {
+      case e: Exception =>
+        logger.error(e.getMessage, e)
+        sys.exit(-1)
+    }
+    sys.exit(0)
+  }
 
   def createCommand(args: Array[String]): Command = {
     jc.setProgramName(scriptName)
