+++ geomesa-tools/src/main/scala/org/locationtech/geomesa/tools/kafka/KafkaRunner.scala
@@ -11,16 +11,4 @@ object KafkaRunner extends Runner {
     new HelpCommand(jc),
     new VersionCommand(jc)
   )
-
-  def main(args: Array[String]): Unit = {
-    val command = createCommand(args)
-    try {
-      command.execute()
-    } catch {
-      case e: Exception =>
-        logger.error(e.getMessage, e)
-        sys.exit(-1)
-    }
-    sys.exit(0)
-  }
 }
